src/requirements.txt
pydantic>=2.9.0
anthropic>=0.40.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

llama-index-core>=0.12.0
llama-index-vector-stores-qdrant>=0.4.0
llama-index-embeddings-fastembed>=0.3.0
fastembed>=0.4.0
llama-index-readers-json>=0.4.0
llama-index-llms-anthropic>=0.6.0
qdrant-client>=1.12.0

langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
chainlit>=2.0.0

crewai>=0.100.0
deepeval>=2.0.0
langfuse>=2.50.0
supabase>=2.0.0
